// frontend/generated/models/PromotionalCampaignRequestDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 8lind8box Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PromotionalCampaignRequestDto
 */
export interface PromotionalCampaignRequestDto {
    /**
     * 
     * @type {number}
     * @memberof PromotionalCampaignRequestDto
     */
    campaignId?: number;
    /**
     * 
     * @type {string}
     * @memberof PromotionalCampaignRequestDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof PromotionalCampaignRequestDto
     */
    description?: string;
    /**
     * 
     * @type {Date}
     * @memberof PromotionalCampaignRequestDto
     */
    startDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PromotionalCampaignRequestDto
     */
    endDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof PromotionalCampaignRequestDto
     */
    discountRate?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PromotionalCampaignRequestDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Array<number>}
     * @memberof PromotionalCampaignRequestDto
     */
    blindBoxIds?: Array<number>;
}

/**
 * Check if a given object implements the PromotionalCampaignRequestDto interface.
 */
export function instanceOfPromotionalCampaignRequestDto(value: object): value is PromotionalCampaignRequestDto {
    return true;
}

export function PromotionalCampaignRequestDtoFromJSON(json: any): PromotionalCampaignRequestDto {
    return PromotionalCampaignRequestDtoFromJSONTyped(json, false);
}

export function PromotionalCampaignRequestDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PromotionalCampaignRequestDto {
    if (json == null) {
        return json;
    }
    return {
        
        'campaignId': json['campaignId'] == null ? undefined : json['campaignId'],
        'title': json['title'] == null ? undefined : json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'startDate': json['startDate'] == null ? undefined : (new Date(json['startDate'])),
        'endDate': json['endDate'] == null ? undefined : (new Date(json['endDate'])),
        'discountRate': json['discountRate'] == null ? undefined : json['discountRate'],
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'blindBoxIds': json['blindBoxIds'] == null ? undefined : json['blindBoxIds'],
    };
}

export function PromotionalCampaignRequestDtoToJSON(json: any): PromotionalCampaignRequestDto {
    return PromotionalCampaignRequestDtoToJSONTyped(json, false);
}

export function PromotionalCampaignRequestDtoToJSONTyped(value?: PromotionalCampaignRequestDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'campaignId': value['campaignId'],
        'title': value['title'],
        'description': value['description'],
        'startDate': value['startDate'] == null ? undefined : ((value['startDate']).toISOString()),
        'endDate': value['endDate'] == null ? undefined : ((value['endDate']).toISOString()),
        'discountRate': value['discountRate'],
        'isVisible': value['isVisible'],
        'blindBoxIds': value['blindBoxIds'],
    };
}