// frontend/generated/models/PromotionCampaignDto.ts
export interface PromotionCampaignDto {
    /**
     * 
     * @type {number}
     * @memberof PromotionCampaignDto
     */
    campaignId?: number;
    /**
     * 
     * @type {string}
     * @memberof PromotionCampaignDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof PromotionCampaignDto
     */
    description?: string;
    /**
     * 
     * @type {Date}
     * @memberof PromotionCampaignDto
     */
    startDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PromotionCampaignDto
     */
    endDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof PromotionCampaignDto
     */
    discountRate?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PromotionCampaignDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Array<BlindBoxCampaignDto>}
     * @memberof PromotionCampaignDto
     */
    blindBoxCampaigns?: Array<BlindBoxCampaignDto>;
    /**
     * 
     * @type {Date}
     * @memberof PromotionCampaignDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PromotionCampaignDto
     */
    updatedAt?: Date;
}